putio-cli/requirements.txt
putio.py==8.7.0
python-dotenv==0.19.2
typer==0.4.0
